corrections

+++ src/headers/tomcrypt_cipher.h
@@ -1018,8 +1018,8 @@ typedef struct {
    ulong32       input[16];
    unsigned char kstream[64];
    unsigned long ksleft;
-   int           rounds;
-   int           status;  /* 0=uninitialized, 1,3=finished setup(), 2=finished setiv() */
+   unsigned long rounds;
+   unsigned long status;  /* 0=uninitialized, 1,3=finished setup(), 2=finished setiv() */
 } salsa20_state;
 
 int salsa20_state_size(void);
@@ -1067,8 +1067,8 @@ typedef struct {
     */
    unsigned char buf[80];
    unsigned      ptr;
-   int           ivlen;
-   int           status;  /* 0=uninitialized, 1,3=finished setup(), 2=finished setiv() */
+   unsigned long ivlen;
+   unsigned long status;  /* 0=uninitialized, 1,3=finished setup(), 2=finished setiv() */
 } sosemanuk_state;
 
 int sosemanuk_state_size(void);
